Include descriptors in workflow replay hash
Use a shared descriptor set builder for workflow artifacts so scope binding replay compares the same artifact hash that revision preparation stores.
