+++ src/platform/Aevatar.GAgentService.Infrastructure/Adapters/WorkflowServiceImplementationAdapter.cs
@@ -1,8 +1,8 @@
 using Aevatar.AI.Abstractions;
 using Aevatar.GAgentService.Abstractions;
+using Aevatar.GAgentService.Abstractions.Services;
 using Aevatar.GAgentService.Core.Ports;
 using Aevatar.Workflow.Application.Abstractions.Runs;
-using Google.Protobuf;
 using Google.Protobuf.Reflection;
 
 namespace Aevatar.GAgentService.Infrastructure.Adapters;
@@ -47,7 +47,7 @@ public sealed class WorkflowServiceImplementationAdapter : IServiceImplementatio
             Identity = request.Spec.Identity.Clone(),
             RevisionId = request.Spec.RevisionId,
             ImplementationKind = ServiceImplementationKind.Workflow,
-            ProtocolDescriptorSet = BuildProtocolDescriptorSet(
+            ProtocolDescriptorSet = ServiceProtocolDescriptorSetBuilder.Build(
                 ChatRequestEvent.Descriptor,
                 ChatResponseEvent.Descriptor),
             Endpoints =
@@ -77,53 +77,4 @@ public sealed class WorkflowServiceImplementationAdapter : IServiceImplementatio
 
     private static string GetTypeUrl(MessageDescriptor descriptor) =>
         $"type.googleapis.com/{descriptor.FullName}";
-
-    private static ByteString BuildProtocolDescriptorSet(params MessageDescriptor[] descriptors)
-    {
-        var files = new Dictionary<string, FileDescriptorProto>(StringComparer.Ordinal);
-        foreach (var descriptor in descriptors)
-            AddFile(descriptor.File, files);
-
-        var descriptorSet = new FileDescriptorSet();
-        descriptorSet.File.Add(files.Values);
-        return descriptorSet.ToByteString();
-    }
-
-    private static void AddFile(
-        FileDescriptor file,
-        IDictionary<string, FileDescriptorProto> files)
-    {
-        if (files.ContainsKey(file.Name))
-            return;
-
-        var proto = file.ToProto();
-        foreach (var dependency in file.Dependencies)
-            AddFile(dependency, files);
-        foreach (var dependency in file.PublicDependencies)
-            AddFile(dependency, files);
-        foreach (var dependencyName in proto.Dependency)
-            AddKnownDependency(dependencyName, files);
-
-        files[file.Name] = proto;
-    }
-
-    private static void AddKnownDependency(
-        string dependencyName,
-        IDictionary<string, FileDescriptorProto> files)
-    {
-        var file = dependencyName switch
-        {
-            "google/protobuf/any.proto" => Google.Protobuf.WellKnownTypes.Any.Descriptor.File,
-            "google/protobuf/duration.proto" => Google.Protobuf.WellKnownTypes.Duration.Descriptor.File,
-            "google/protobuf/empty.proto" => Google.Protobuf.WellKnownTypes.Empty.Descriptor.File,
-            "google/protobuf/field_mask.proto" => Google.Protobuf.WellKnownTypes.FieldMask.Descriptor.File,
-            "google/protobuf/struct.proto" => Google.Protobuf.WellKnownTypes.Struct.Descriptor.File,
-            "google/protobuf/timestamp.proto" => Google.Protobuf.WellKnownTypes.Timestamp.Descriptor.File,
-            "google/protobuf/wrappers.proto" => Google.Protobuf.WellKnownTypes.StringValue.Descriptor.File,
-            _ => null,
-        };
-
-        if (file != null)
-            AddFile(file, files);
-    }
 }
